advanced/multi_region_s3_crr_kms_cmk/requirements.txt
aws-cdk.core==1.137.0
-e constructs/multi_region_s3_crr_kms_cmk_source
-e constructs/multi_region_s3_crr_kms_cmk_target
